+++ backend/src/services/handleErrors.js
@@ -1,7 +1,7 @@
 const errorHandler = (err) => {
   if (err instanceof Error) {
     console.log(err);
-    throw { code: 500, message: err };
+    throw { code: 500, message: err.message };
   } else if (err.code) {
     throw err;
   } else {